Moved related code to finalEngine.cpp and viewportEngine.cpp
